+++ babel.config.js
@@ -1,3 +1,22 @@
-module.exports = () => ({
-  babelrcRoots: ['.', 'packages/*'],
-});
+module.exports = (api) => {
+  api.cache(false);
+  api.assertVersion('^7.4.4');
+
+  const presets = [
+    '@babel/preset-env',
+    '@babel/preset-react',
+  ];
+
+  const plugins = [
+    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-proposal-object-rest-spread',
+    '@babel/plugin-transform-async-to-generator',
+    '@babel/plugin-transform-object-assign',
+    '@babel/plugin-transform-runtime',
+  ];
+
+  return {
+    presets,
+    plugins,
+  };
+};